Move navigation to modules
